refactor: rename backend to gateway
