+++ apps/gateway/.dockerignore
